requirements.txt
python-dotenv==0.21.1
py7zr==0.21.0